refactor visitors to adher to better design pattern and separation of concerns
