+++ Salon/src/directory_processor/repo_readme_generator.py
@@ -7,13 +7,16 @@ and if so, creates or updates it using a chosen summarization model.
 import os
 from datetime import datetime
 from typing import Optional
-from .models.base import AIModel
+from ..models.base import AIModel
 from CommonPy.src.request_utilities import request_timeout
 
-from .directory_visitor_base import DirectoryVisitor, DirectoryData
-from .models.model_factory import create_model
+from .base import DirectoryProcessor
+from ..types.directory_data import DirectoryData
+from ..models.model_factory import create_model
+from ..core.file_handler import FileHandler
+from ..prompts.prompts import CODE_SUMMARIZER_PERSONA_INTRO
 
-class DirectoryVisitorForReadme(DirectoryVisitor):
+class ReadmeGenerator(DirectoryProcessor):
     """
     Creates or updates 'ReadMe.Salon.md' by summarizing the directory's
     source files if they are newer than the existing readme.
@@ -21,20 +24,18 @@ class DirectoryVisitorForReadme(DirectoryVisitor):
     SUMMARY_FILENAME = 'ReadMe.Salon.md'
 
     def __init__(self, model_type: str = "braid_api", priority: int = 1) -> None:
-        """
-        :param model_type: Summarisation model type: 'braid_api' or 'local_gemini'.
-        """
         super().__init__(priority=priority)
         print(f"Creating model: {model_type}")
         self.driver: AIModel = create_model(model_type.lower())
+        self.file_handler = FileHandler()  # Instantiate FileHandler
 
     def summarise_code(self, code: str) -> Optional[str]:
         """
         Summarize source code text using the configured model driver.
         """
-        return self.driver.generate_content(code, persona="CodeSummariser", 
-                                      persona_intro="You are an AI assistant that summarizes code to help explain it to new developers.",
-                                      length_in_words=100)
+        return self.driver.generate_content(code, persona="CodeSummariser",
+                                            persona_intro= CODE_SUMMARIZER_PERSONA_INTRO,
+                                            length_in_words=100)
 
     def visit(self, directory_data: DirectoryData) -> None:
         """
@@ -60,25 +61,17 @@ class DirectoryVisitorForReadme(DirectoryVisitor):
 
         new_readme = []
         for src_file in directory_data.source_files:
-            try:
-                with open(src_file, 'r', encoding='utf-8') as f:
-                    code = f.read().rstrip()
-                if len(code) > 250:  # Summarize only if file is somewhat large
-                    summary = self.summarise_code(code)
-                    if summary:
-                        new_readme.append(f"**{src_file.name}**\n\n{summary}\n")
-            except Exception as e:
-                print(f"Failed to summarize {src_file}: {e}")
+            code = self.file_handler.read_file(src_file) # Use FileHandler to read files
+            if code and len(code) > 250:  # Summarize only if file is somewhat large
+                summary = self.summarise_code(code)
+                if summary:
+                    new_readme.append(f"**{src_file.name}**\n\n{summary}\n")
 
         if new_readme:
-            # Append "Generated by Salon..." comment with preceding newline
             salon_comment = (
                 f"\nGenerated by Salon from Braid Technologies, "
                 f"{datetime.now().strftime('%d/%m/%Y')}"
             )
-
-            with open(readme_path, 'w', encoding='utf-8') as f:
-                f.write("\n".join(new_readme))
-                f.write(salon_comment)
-
+            content = "\n".join(new_readme) + salon_comment
+            self.file_handler.write_file_version(directory_data.path, self.SUMMARY_FILENAME, content) # Use FileHandler to write files
             print(f"Created/Updated {readme_path}")
